requirements.txt
transformers
sentence-transformers
faiss-cpu
PyMuPDF
gradio
torch
